Adding some more neuroshare supports. But Neuroshare is not optimal!
