+++ circus/files/plexon.py
@@ -1,0 +1,8 @@
+import h5py, numpy, re, sys
+from neuroshare import NeuroShareFile
+import neuroshare as ns
+
+class PlexonFile(NeuroShareFile):
+
+    description    = "plexon"    
+    extension      = [".plx"]
